backend/requirements.txt
flask==3.1.0
flask-cors==5.0.1
flask-limiter==3.5.0
python-dotenv==1.0.0
yfinance==0.2.51
numpy==2.2.3
pandas==2.2.3
